+++ packages/spor-accordion-react-native/src/Expandable.tsx
