+++ packages/spor-accordion-react-native/src/index.tsx
@@ -1,1 +1,1 @@
-export * from "./Accordion";
+export * from "./Expandable";
